Remove obsolete `allow(warnings)` attributes (#373)
We were for a time allowing warnings on two sections of code due to
continued use of `...` pattern syntax.  We had deferred the change to
`..=` syntax so as to support earlier Rust versions.

However, in commit `6059be7a99a264abf9afb8e6b8461413b972e892` we
changed over to `..=` syntax.  The `allow(warnings)` attributes and
the related comments are now obsolete, and we remove them in this
commit.